+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.20.0-dev
 
+- Soft deprecations (no warnings)
+    - The query builder has been removed from documentation and will be eventually removed completely. This will be done because of the current limitations of the builder implementation and the InfluxDB move from InfluxQL to Flux as the query language of choice
+
 - Deprecations
     - Accessing the system environment by configuring `{:system, var}` or `{:system, var, default}` will now result in a `Logger.info/1` message and will stop working in a future release