+++ CHANGELOG.md
@@ -12,7 +12,7 @@
     - Accessing the system environment by configuring `{:system, var}` or `{:system, var, default}` will now result in a `Logger.info/1` message and will stop working in a future release
 
 - Backwards incompatible changes
-    - Support for the InfluxDB version earlier than `1.4.x` is no longer guaranteed
+    - Support for the InfluxDB versions earlier than `1.4.x` is no longer guaranteed
 
 ## v0.19.0 (2019-01-24)